Extract security classes to common library
