+++ src/main/java/uk/nhs/cdss/config/SecurityConfig.java
@@ -1,15 +1,17 @@
 package uk.nhs.cdss.config;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import uk.nhs.cdss.security.JWTFilter;
+import uk.nhs.cactus.common.security.JWTFilter;
 
 @Configuration
 @RequiredArgsConstructor
+@ComponentScan("uk.nhs.cactus.common.security")
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
   private final JWTFilter jwtFilter;
