+++ src/main/java/uk/nhs/cdss/service/GenericResourceLocator.java
@@ -8,7 +8,7 @@ import org.hl7.fhir.dstu3.model.Reference;
 import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.instance.model.api.IIdType;
 import org.springframework.stereotype.Service;
-import uk.nhs.cdss.security.AuthenticatedFhirClientFactory;
+import uk.nhs.cactus.common.security.AuthenticatedFhirClientFactory;
 import uk.nhs.cdss.util.RetryUtils;
 
 @Service
